Improve examples
Conflicts:
	examples/http-example/main.go
	examples/http-jwks-example/main.go
