+++ examples/http-example/README.md
@@ -4,6 +4,9 @@ This is an example of how to use the http middleware.
 
 # Using it
 
-To try this out, first install all dependencies with `go mod download` and then run `go run main.go` to start the app.
+To try this out:
 
-* Call `http://localhost:3000` with a JWT signed with `My Secret` (you can use [jwt.io](https://jwt.io/) for this) to get a response back.
+* Install all dependencies with `go mod vendor`.
+* Run `go run main.go` to start the app.
+* Use [jwt.io](https://jwt.io/) to generate a JWT signed with the HS256 algorithm and `secret`.
+* Call `http://localhost:3000` with the JWT to get a response back.
