+++ examples/http-jwks-example/main.go
@@ -7,18 +7,14 @@ import (
 	"net/url"
 	"time"
 
-	"gopkg.in/square/go-jose.v2"
-	"gopkg.in/square/go-jose.v2/jwt"
-
-	"github.com/auth0/go-jwt-middleware/validate/josev2"
-
-	"github.com/auth0/go-jwt-middleware"
 	"github.com/auth0/go-jwt-middleware/jwks"
 	"github.com/auth0/go-jwt-middleware/validator"
+
+	"github.com/auth0/go-jwt-middleware"
 )
 
 var handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-	claims := r.Context().Value(jwtmiddleware.ContextKey{}).(*josev2.UserContext)
+	claims := r.Context().Value(jwtmiddleware.ContextKey{}).(*validator.ValidatedClaims)
 
 	payload, err := json.Marshal(claims)
 	if err != nil {
@@ -31,26 +27,26 @@ var handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 })
 
 func main() {
-	issuerURL, err := url.Parse("https://<your tenant domain>")
+	issuerURL, err := url.Parse("https://<your tenant domain>/")
 	if err != nil {
 		log.Fatalf("failed to parse the issuer url: %v", err)
 	}
 
 	provider := jwks.NewCachingProvider(issuerURL, 5*time.Minute)
 
-	// Set up the josev2 validator.
-	validator, err := josev2.New(
+	// Set up the validator.
+	jwtValidator, err := validator.New(
 		provider.KeyFunc,
-		jose.RS256,
-		"josev2-example",
-		jwt.Audience{},
+		"RS256",
+		issuerURL.String(),
+		[]string{"<your api identifier>"},
 	)
 	if err != nil {
-		log.Fatalf("failed to set up the josev2 validator: %v", err)
+		log.Fatalf("failed to set up the validator: %v", err)
 	}
 
 	// Set up the middleware.
-	middleware := jwtmiddleware.New(validator.ValidateToken)
+	middleware := jwtmiddleware.New(jwtValidator.ValidateToken)
 
 	http.ListenAndServe("0.0.0.0:3000", middleware.CheckJWT(handler))
 }
