+++ validator/doc.go
@@ -1,5 +1,5 @@
 /*
-Package josev2 contains an implementation of jwtmiddleware.ValidateToken using
+Package validator contains an implementation of jwtmiddleware.ValidateToken using
 the Square go-jose package version 2.
 
 The implementation handles some nuances around JWTs and supports:
